add CheckMark to FilterChoice refactor extension to Helpers
