+++ Tracker/Helpers/UICollectionExtention.swift
@@ -1,0 +1,42 @@
+//
+//  UICollectionExtention.swift
+//  Tracker
+//
+//  Created by Irina Deeva on 18/04/24.
+//
+
+import UIKit
+
+extension UICollectionView {
+    func setEmptyMessage(message: String, image: String) {
+        let emptyView = UIView(frame: CGRect(x: 0, y: 0, width: self.bounds.size.width, height: self.bounds.size.height))
+        emptyView.backgroundColor = .ypWhite
+        emptyView.sizeToFit()
+
+        let imageStub = UIImageView()
+        imageStub.image = UIImage(named: image) ?? UIImage()
+        emptyView.addSubview(imageStub)
+
+        let labelStub = UILabel()
+        labelStub.text = message
+        labelStub.font = .systemFont(ofSize: 12, weight: .medium)
+        labelStub.textColor = .ypBlackDay
+        emptyView.addSubview(labelStub)
+
+        imageStub.translatesAutoresizingMaskIntoConstraints = false
+        labelStub.translatesAutoresizingMaskIntoConstraints = false
+
+        NSLayoutConstraint.activate([
+            imageStub.centerXAnchor.constraint(equalTo: emptyView.centerXAnchor),
+            imageStub.centerYAnchor.constraint(equalTo: emptyView.centerYAnchor),
+            labelStub.centerXAnchor.constraint(equalTo: emptyView.centerXAnchor),
+            labelStub.topAnchor.constraint(equalTo: imageStub.bottomAnchor, constant: 8)
+        ])
+
+        self.backgroundView = emptyView
+    }
+
+    func restore() {
+        self.backgroundView = nil
+    }
+}
