+++ Tracker/UI/TrackerCollection/TrackersViewController.swift
@@ -18,7 +18,7 @@ final class TrackersViewController: UIViewController {
 
     private var filteredCategories: [TrackerCategory] = []
     private var completedTrackers: Set<TrackerRecord> = []
-    private var selectedFilter: Filter = .all
+    private var selectedFilter: Filter?
 
     private var isSearchBarEmpty: Bool {
         return searchController.searchBar.text?.isEmpty ?? true
@@ -342,40 +342,6 @@ extension TrackersViewController: UISearchResultsUpdating {
     }
 }
 
-extension UICollectionView {
-    func setEmptyMessage(message: String, image: String) {
-        let emptyView = UIView(frame: CGRect(x: 0, y: 0, width: self.bounds.size.width, height: self.bounds.size.height))
-        emptyView.backgroundColor = .ypWhite
-        emptyView.sizeToFit()
-
-        let imageStub = UIImageView()
-        imageStub.image = UIImage(named: image) ?? UIImage()
-        emptyView.addSubview(imageStub)
-
-        let labelStub = UILabel()
-        labelStub.text = message
-        labelStub.font = .systemFont(ofSize: 12, weight: .medium)
-        labelStub.textColor = .ypBlackDay
-        emptyView.addSubview(labelStub)
-
-        imageStub.translatesAutoresizingMaskIntoConstraints = false
-        labelStub.translatesAutoresizingMaskIntoConstraints = false
-
-        NSLayoutConstraint.activate([
-            imageStub.centerXAnchor.constraint(equalTo: emptyView.centerXAnchor),
-            imageStub.centerYAnchor.constraint(equalTo: emptyView.centerYAnchor),
-            labelStub.centerXAnchor.constraint(equalTo: emptyView.centerXAnchor),
-            labelStub.topAnchor.constraint(equalTo: imageStub.bottomAnchor, constant: 8)
-        ])
-
-        self.backgroundView = emptyView
-    }
-
-    func restore() {
-        self.backgroundView = nil
-    }
-}
-
 extension TrackersViewController: TrackerCellButtonDelegate {
     func didTapButtonInCell(_ cell: TrackerCell) {
 
@@ -424,12 +390,6 @@ extension TrackersViewController: ChoiceTrackerDelegate {
 
 extension TrackersViewController: FilterChoiceDelegate {
     func didDoneTapped(_ filter: Filter) {
-        
-    }
-}
-
-extension Date {
-    var startOfDay: Date {
-        return Calendar.current.startOfDay(for: self)
+        selectedFilter = filter
     }
 }
